move shared scaler routines to shared module
