+++ external/fv3fit/fv3fit/pytorch/graph/train.py
@@ -8,17 +8,23 @@ from fv3fit.pytorch.graph.network import GraphNetwork, GraphNetworkConfig
 from fv3fit.pytorch.loss import LossConfig
 from fv3fit.pytorch.optimizer import OptimizerConfig
 from fv3fit.pytorch.training_loop import AutoregressiveTrainingConfig
-from fv3fit._shared.scaler import StandardScaler
+from fv3fit._shared.scaler import (
+    get_standard_scaler_mapping,
+    get_mapping_standard_scale_func,
+    StandardScaler,
+)
 from ..system import DEVICE
 
 from fv3fit._shared import register_training_function
 from typing import (
     Callable,
+    Hashable,
     List,
     Optional,
     Sequence,
     Set,
     Mapping,
+    cast,
 )
 from fv3fit.tfdataset import select_keys, ensure_nd, apply_to_mapping
 
@@ -53,27 +59,6 @@ class GraphHyperparameters(Hyperparameters):
         return set(self.state_variables)
 
 
-def get_scalers(sample: Mapping[str, np.ndarray]):
-    scalers = {}
-    for name, array in sample.items():
-        s = StandardScaler(n_sample_dims=5)
-        s.fit(array)
-        scalers[name] = s
-    return scalers
-
-
-def get_mapping_scale_func(
-    scalers: Mapping[str, StandardScaler]
-) -> Callable[[Mapping[str, np.ndarray]], Mapping[str, np.ndarray]]:
-    def scale(data: Mapping[str, np.ndarray]):
-        output = {**data}
-        for name, array in data.items():
-            output[name] = scalers[name].normalize(array)
-        return output
-
-    return scale
-
-
 # TODO: Still have to handle forcing
 
 
@@ -93,13 +78,12 @@ def train_graph_model(
         validation_batches: validation data, as a dataset of Mapping[str, tf.Tensor]
             where each tensor has dimensions [batch, time, tile, x, y(, z)]
     """
-    train_batches = train_batches.map(apply_to_mapping(ensure_nd(6)))
     sample = next(
         iter(train_batches.unbatch().batch(hyperparameters.normalization_fit_samples))
     )
 
-    scalers = get_scalers(sample)
-    mapping_scale_func = get_mapping_scale_func(scalers)
+    scalers = get_standard_scaler_mapping(sample)
+    mapping_scale_func = get_mapping_standard_scale_func(scalers)
 
     get_state = curry(get_Xy_dataset)(
         state_variables=hyperparameters.state_variables,
@@ -130,7 +114,7 @@ def train_graph_model(
     predictor = PytorchAutoregressor(
         state_variables=hyperparameters.state_variables,
         model=train_model,
-        scalers=scalers,
+        scalers=cast(Mapping[Hashable, StandardScaler], scalers),
     )
     return predictor
